requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.4.1=pyhd8ed1ab_0
awscli=1.33.7=pypi_0
boto3=1.34.125=pypi_0
botocore=1.34.125=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.6.2=hbcca054_0
certifi=2024.6.2=pypi_0
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pyhd8ed1ab_0
debugpy=1.6.7=py312h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
distro=1.9.0=pypi_0
docutils=0.16=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h6a678d5_0
greenlet=3.0.3=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
install=1.3.5=pypi_0
ipykernel=6.29.4=pyh3099207_0
ipython=8.25.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jmespath=1.0.1=pypi_0
joblib=1.4.2=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.5.0=py312h06a4308_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=13.2.0=h77fa898_8
libgomp=13.2.0=h77fa898_8
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
numpy=2.0.0=pypi_0
openssl=3.3.1=h4ab18f5_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pip=24.0=py312h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=5.9.0=py312h5eee18b_0
psycopg2-binary=2.9.9=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.6.0=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
python=3.12.3=h996f2a0_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py312h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
rsa=4.7.2=pypi_0
s3transfer=0.10.1=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
setuptools=69.5.1=py312h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlalchemy=2.0.30=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
tabula-py=2.9.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
tornado=6.3.3=py312h5eee18b_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
